% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{all_cpp_3}
\alias{all_cpp_3}
\title{Are all Elements in a Vector "TRUE"? (Form 3)}
\usage{
all_cpp_3(x)
}

\arguments{
\item{x}{logical vector.}
}

\description{
Loop over a vector and returns "true" if all the elements are "true".
}

