% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{any_cpp}
\alias{any_cpp}
\title{Is Any Element in a Logical Vector "TRUE"}
\usage{
any_cpp(x)
}

\arguments{
\item{x}{logical vector.}
}

\description{
Loop over a vector and returns "true" when a "true" element is found in the sequence.
}

