% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in 05-logical-functions.R
\name{any_r}
\alias{any_r}
\title{Is Any Element in a Logical Vector "TRUE"}
\usage{
any_r(x)
}

\arguments{
\item{x}{logical vector.}
}

\description{
Loop over a vector and returns "true" when a "true" element is found in the sequence.
}

