% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4rvignettes-package.R
\name{cpp4rvignettes-package}
\alias{cpp4rvignettes-package}
\title{Examples Used in the Vignettes}
\description{
Covers different examples using 'R' and 'C++' code. The functions provided here
 serve as examples and are used to run benchmarks.
}

\keyword{internal}
