% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{cumprod_cpp}
\alias{cumprod_cpp}
\title{Cumulative Product}
\usage{
cumprod_cpp(x)
}

\arguments{
\item{x}{doubles (numeric) vector.}
}

\description{
Returns a vector of doubles containing the cumulative product of a vector of doubles.
}

