% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in 06-rolling-functions.R
\name{cumprod_r}
\alias{cumprod_r}
\title{Cumulative Product}
\usage{
cumprod_r(x)
}

\arguments{
\item{x}{doubles (numeric) vector.}
}

\description{
Returns a vector of doubles containing the cumulative product of a vector of doubles.
}

