% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{cumsum_cpp}
\alias{cumsum_cpp}
\title{Cumulative Sum}
\usage{
cumsum_cpp(x)
}

\arguments{
\item{x}{doubles (numeric) vector.}
}

\description{
Returns a vector of doubles containing the cumulative sum of a vector of doubles.
}

