% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in 06-rolling-functions.R
\name{cumsum_r}
\alias{cumsum_r}
\title{Cumulative Sum}
\usage{
cumsum_r(x)
}

\arguments{
\item{x}{doubles (numeric) vector.}
}

\description{
Returns a vector of doubles containing the cumulative sum of a vector of doubles.
}

