% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{range_cpp}
\alias{range_cpp}
\title{Cumulative Product}
\usage{
range_cpp(x)
}

\arguments{
\item{x}{doubles (numeric) vector.}
}

\description{
Returns a vector of doubles containing the minimum and maximum of a vector of doubles.
}

