% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{which_cpp}
\alias{which_cpp}
\title{Which Elements in a Logical Vector are "TRUE"}
\usage{
which_cpp(x)
}

\arguments{
\item{x}{logical vector.}
}

\description{
Loop over a vector and returns the positions that are "true". Returns a vector of
  integers.
}

