% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in 05-logical-functions.R
\name{which_r}
\alias{which_r}
\title{Which Elements in a Logical Vector are "TRUE"}
\usage{
which_r(x)
}

\arguments{
\item{x}{logical vector.}
}

\description{
Loop over a vector and returns the positions that are "true". Returns a vector of
 integers.
}

